TrainFutures: Move DestroyableThread to caikit.core
This will be used in the local train backend
