--- caikit/runtime/types/thread_destroyed_exception.py
@@ -1,24 +1,0 @@
-# Copyright The Caikit Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-class ThreadDestroyedException(RuntimeError):
-    """Exception raised inside a DestroyableThread when it is destroyed by the thread managing
-    its lifecycle."""
-
-    def __init__(self):
-        super().__init__(
-            "Work thread intentionally destroyed by its lifecycle manager. "
-            "This exception was not raised by the code running in this thread."
-        )
